fix: move get workflow type to service
